refactor: controller endpoints now recieve DTO through params
