+++ src/main/java/com/laversi/schoolapi/controllers/personController/PersonController.java
@@ -3,7 +3,7 @@ package com.laversi.schoolapi.controllers.personController;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 
-import com.laversi.schoolapi.model.personDTO.PersonDTO;
+import com.laversi.schoolapi.model.personDTO.*;
 import com.laversi.schoolapi.model.personEntity.PersonEntity;
 import com.laversi.schoolapi.repositories.PersonRepository;
 
@@ -40,23 +40,39 @@ public class PersonController {
         return ResponseEntity.ok(convertToDto(savedPerson));
     }
     @PutMapping("/{id}")
-    public ResponseEntity<PersonDTO> updatePerson(@PathVariable String id, @RequestBody PersonDTO person) {
+    public ResponseEntity<PersonDTO> updatePerson(@PathVariable String id, @RequestBody PersonPutDTO person) {
         Optional<PersonEntity> optionalPerson = repo.findById(id);
-        if (optionalPerson.isEmpty()) {return ResponseEntity.notFound().build();}
+        if (optionalPerson.isEmpty()) {
+            return ResponseEntity.notFound().build();
+        }
     
         PersonEntity existingPerson = optionalPerson.get();
-        if (person.name() != null) {existingPerson.setName(person.name());}
-        if (person.email() != null) {existingPerson.setEmail(person.email());}
-        if (person.password() != null) {existingPerson.setPassword(person.password());}
-        if (person.birthDate() != null) {existingPerson.setBirthDate(person.birthDate());}
-        if (person.motherName() != null) {existingPerson.setMotherName(person.motherName());}
-        if (person.fatherName() != null) {existingPerson.setFatherName(person.fatherName());}
-        if (person.telephone() != null) {existingPerson.setTelephone(person.telephone());}    
+        if (person.name() != null) {
+            existingPerson.setName(person.name());
+        }
+        if (person.email() != null) {
+            existingPerson.setEmail(person.email());
+        }
+        if (person.password() != null) {
+            existingPerson.setPassword(person.password());
+        }
+        if (person.birthDate() != null) {
+            existingPerson.setBirthDate(person.birthDate());
+        }
+        if (person.motherName() != null) {
+            existingPerson.setMotherName(person.motherName());
+        }
+        if (person.fatherName() != null) {
+            existingPerson.setFatherName(person.fatherName());
+        }
+        if (person.telephone() != null) {
+            existingPerson.setTelephone(person.telephone());
+        }
+    
         PersonEntity updatedPerson = repo.save(existingPerson);
-
         return ResponseEntity.ok(convertToDto(updatedPerson));
     }
-
+    
     @DeleteMapping("/{id}")
     public ResponseEntity<Void> deletePerson(@PathVariable String id) {
         repo.deleteById(id);
