+++ src/main/java/com/laversi/schoolapi/model/personDTO/PersonPutDTO.Java
@@ -1,4 +1,5 @@
-package com.laversi.schoolapi.model.personPutDTO;
+package com.laversi.schoolapi.model.personDTO;
+import java.time.LocalDate;
 
 public record PersonPutDTO(
         String name,
